Switch to glide, update to latest k8s.
Manual merges were done in the k2s/kube2sky.go file. Type changes were
required in a few other locations.

+++ proxy.go
@@ -15,7 +15,7 @@ const (
 )
 
 var (
-	MasqueradeBit = 14
+	MasqueradeBit = int32(14)
 	ProxyStop     chan struct{}
 )
 
@@ -40,7 +40,7 @@ func StartProxyServer() {
 	config.Mode = componentconfig.ProxyModeUserspace
 
 	// defaults
-	oom := qos.KubeProxyOOMScoreAdj
+	oom := int32(qos.KubeProxyOOMScoreAdj)
 	config.OOMScoreAdj = &oom
 	config.IPTablesMasqueradeBit = &MasqueradeBit
